retina/requirements.txt
asgiref==3.7.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.10.0
Django==4.2.4
djangorestframework==3.14.0
fonttools==4.42.1
idna==2.10
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==9.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.4
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
supervision==0.14.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
wget==3.2
